refactor: spell out translationengleiche module
